refactor(agent-toolkit): co-locate workflow GraphQL files with their tools
Move shared board-automations and workflow mutation files into their
respective tool directories, updating imports accordingly.
