+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workflows-tools/delete-workflow/delete-workflow.graphql.dev.ts
@@ -1,0 +1,9 @@
+import { gql } from 'graphql-request';
+
+export const deleteLiveWorkflowMutation = gql`
+  mutation deleteLiveWorkflow($id: ID!) {
+    delete_live_workflow(id: $id) {
+      is_success
+    }
+  }
+`;
